Cleaning up of code in model.
Creation of a database module for DbSchemaM and DbHelperM.
Creation of a generator for the update steps.
Encapsulation of all needed into the DbHelperM class.
